requirements.txt
ffmpeg-python==0.2.0
langchain==0.0.300
openai-whisper==20231117
streamlit==1.29.0
openai==1.5.0